+++ eval/eval_detection.py
@@ -273,8 +273,8 @@ def compute_average_precision_detection(ground_truth, prediction, tiou_threshold
             if fp[tidx, idx] == 0 and tp[tidx, idx] == 0:
                 fp[tidx, idx] = 1
 
-    tp_cumsum = np.cumsum(tp, axis=1).astype(np.float)
-    fp_cumsum = np.cumsum(fp, axis=1).astype(np.float)
+    tp_cumsum = np.cumsum(tp, axis=1).astype(np.float32)
+    fp_cumsum = np.cumsum(fp, axis=1).astype(np.float32)
     recall_cumsum = tp_cumsum / npos
     precision_cumsum = tp_cumsum / (tp_cumsum + fp_cumsum)
     # f2score_cumsum = (5*recall_cumsum*precision_cumsum)/(precision_cumsum+4*recall_cumsum+1e-8)
